first design ideas by CJR
This design used the HTML and css from the actual housing lottery pages
as a home page. Links were added under the left sidebar menu to possible
designs for revamping the selection of a room.
The 3 links did a course implementation of the following:
1) Bar chart with roll-over tooltips giving building, room type, and
rooms left for selection
2) Zoomable bubble chart of campus areas, building in areas, room types
with available rooms
3) Same as (2) with a tree map.