Some possible baseline files
This is the HTML, css and data files that I used for my initial design.
Modify at will.
I am not including StudentSite.css - this file is copyrighted. If there
is missing "stuff" let me know and I will add it to the project2.css.
Otherwise we just use it for the demo